requirements.txt
asyncio==3.4.3
aiogram==3.3.0
openai==1.11.1
mysql-connector-python==8.3.0
python-dateutil==2.8.2
asyncio==3.4.3
config==0.5.1
python-dotenv==1.0.1
telethon==1.34.0
aiomysql==0.2.0
FastAPI==0.109.2
loguru==0.7.2
